requirements.txt
python-docx
PyPDF2
pandas
openpyxl